+++ scripts/deploy.sh
@@ -7,4 +7,5 @@ poetry build
 LATEST_VERSION=$(basename $(ls -rt dist/*.whl | tail -n 1))
 
 scp dist/$LATEST_VERSION pi@ubo-development-pod:/tmp/
-ssh pi@ubo-development-pod "source ubo-app/bin/activate && pip install --upgrade --force-reinstall --no-deps /tmp/$LATEST_VERSION && (killall ubo -9 || true) && ubo"
+test "$deps" == "True" && ssh pi@ubo-development-pod "source ubo-app/bin/activate && pip install --upgrade /tmp/$LATEST_VERSION[default]"
+ssh pi@ubo-development-pod "source \$HOME/.profile && source /etc/profile && source ubo-app/bin/activate && pip install --upgrade --force-reinstal --no-deps /tmp/$LATEST_VERSION && (sudo killall ubo -9 || true) && sudo \$(which ubo)"