+++ ubo_app/services/0-keyboard/__init__.py
@@ -1,0 +1,2 @@
+ubo_service_id = 'keyboard'
+ubo_service_name = 'Keyboard'