+++ ubo_app/services/0-keypad/__init__.py
@@ -1,0 +1,2 @@
+ubo_service_id = 'keypad'
+ubo_service_name = 'Keypad'