+++ ubo_app/services/0-sound/__init__.py
@@ -1,0 +1,2 @@
+ubo_service_id = 'sound'
+ubo_service_name = 'Sound'